Formatted a little bit on the VideoPlayerRecyclerView.java
added video_player_placeholder.png, but not used just yet.

Working on adding a placehoder image for with videos are loading, and am also trying to add a thumbnail of the first frame of the video. It is a bit "laggy", need to work on this more. The code for it is currently commented out. Videos work fine.

Created started layout for uploading a video directly from YouTube.