leaving the breakpoints on purpose in case I must debug for uploading youtube links
extractYoutubeUrl extracts the appropriate mp4 link from a youtube link
https://github.com/HaarigerHarald/android-youtubeExtractor

seems to work with modern-ish youtube links, but a 2 second video from 12 years ago was giving trouble. Need to test how recent the videos have to be.

This version properly uploads the decoded youtube url to the Parse backend and properly loads the video into the feed (it also plays)