=- need to implement progress for loading the picture/uploading it
added proper icons for comments, saving, and liking.
saving and liking has nice animations used from LikeButton library.

Liking a post now stores the post's objectID is a string that belongs to the User class.
For the most part is appears to work correctly and it keeps track of a the posts a user has liked/unliked.
However, my implementation im sure if very inneficient. Must improve on it.